new `parametersrecord.py`; ranamed `Tensor`->`Array`; more
